+++ test/api_get_save_product_test.dart
@@ -150,8 +150,8 @@ void main() {
       expect(product.additives!.names[4], 'E950');
 
       expect(product.images, isNotNull);
-      expect(product.images!, hasLength(10));
-      expect(product.getRawImages(), hasLength(6));
+      expect(product.images!, hasLength(7));
+      expect(product.getRawImages(), hasLength(3));
       expect(product.getMainImages(), hasLength(4));
       expect(product.countries, 'Frankreich,Deutschland');
     });